+++ src/main/java/com/tmdigital/gestiondestock/services/impl/SupplierServiceImpl.java
@@ -8,6 +8,8 @@ import org.springframework.stereotype.Service;
 import com.tmdigital.gestiondestock.dto.SupplierDto;
 import com.tmdigital.gestiondestock.exception.ErrorCodes;
 import com.tmdigital.gestiondestock.exception.InvalidEntityException;
+import com.tmdigital.gestiondestock.model.OrderSupplier;
+import com.tmdigital.gestiondestock.repository.OrderSupplierRepository;
 import com.tmdigital.gestiondestock.repository.SupplierRepository;
 import com.tmdigital.gestiondestock.services.SupplierService;
 import com.tmdigital.gestiondestock.validator.SupplierValidator;
@@ -19,9 +21,11 @@ import lombok.extern.slf4j.Slf4j;
 public class SupplierServiceImpl implements SupplierService {
 
     private final SupplierRepository supplierRepository;
+    private final OrderSupplierRepository orderSupplierRepository;
 
-    public SupplierServiceImpl(SupplierRepository supplierRepository) {
+    public SupplierServiceImpl(SupplierRepository supplierRepository, OrderSupplierRepository orderSupplierRepository) {
         this.supplierRepository = supplierRepository;
+        this.orderSupplierRepository = orderSupplierRepository;
     }
 
     @Override
@@ -75,6 +79,14 @@ public class SupplierServiceImpl implements SupplierService {
             log.error("L'identifiant est nul");
             return;
         }
+
+        List<OrderSupplier> ordersSupplier = orderSupplierRepository.findAllBySupplierId(id);
+
+        if (!ordersSupplier.isEmpty()) {
+            log.warn("Le fournisseur avec l'ID {} ne peut pas être supprimé car il est lié à des commandes", id);
+            throw new InvalidEntityException("Impossible de supprimer le fournisseur car il est lié à des commandes", ErrorCodes.SUPPLIER_ALREADY_IN_USE);
+        }
+        
         supplierRepository.deleteById(id);
     }
